IS-1959: Migrate behandler radiogroup to Aksel and inline
